Add an optional timeout value for local editing.
The idea here is that, if you have your client running long-term on a
remote server, you will accumulate a lot of editor files. With the
timeout enabled, it assumes you're done working on the file after
<timeout> has transpired and it's removed from monitoring for changes
and the intermediary file is deleted.

This isn't the greatest solution, but it works until the next idea is
implemented.